add inline documentation for the curl command-line tool
Curl requires a higher threshold of data (4096 bytes) to start
streaming content from the server. The current example is optimized for
a browser-expierince, but it can be tested at the command-line with
curl. To do so only requires one extra flag to be set in order for it
to work: `--no-buffer`