Make Falcon::Adapters::Input#read behave like IO#read
Falcon::Adapters::Input#read at the moment doesn't respect the output
buffer argument when length is given. This makes IO.copy_stream not
behave correctly when a Falcon::Adapters::Input object is used as the
source argument.

This modifies Falcon::Adapters::Input#read to append retrieved data to
the output buffer. The output buffer is now also cleared beforehand,
just like in IO#read.

We also return nil when EOF was reached and length argument was given,
like IO#read. When length is 0, IO#read always returns an empty string,
so we handle that case as well (even though it's rather peculiar).