Deallocate chunk of string after using
This reduces the total memory allocated of
`Falcon::Adapter::Input#read(size, buffer)` from 6x to 5x the Rack input
size.

The buffer string is used for example when using the "rack.input" object
as the source argument for `IO.copy_stream`, or in
Rack::Multipart::Parser (in the next Rack version), or one could use the
buffer string manually.

It was necessary to also switch from calling String#replace on the
buffer string to String#clear + String#<<, due to the Ruby ticket linked
in the comment. Otherwise clearing the chunk doesn't have any effect on
the memory allocated.